Update README.md
From reading the README, I think the `projectPath` setting is mistaken for the `httpPath` one, since no explanation for the former is found anywhere. This patch changes the references accordingly.

If there is, in fact, a `projectPath` setting in grunt-contrib-compass, it should also be explained in this file, because otherwise the explanations relying on it remain incomplete.